+++ test/OpenTelemetry.Instrumentation.SqlClient.Tests/SqlEventSourceTests.netfx.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="SqlEventSourceTests.netfx.cs" company="OpenTelemetry Authors">
+// <copyright file="SqlEventSourceTests.netfx.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");