+++ test/OpenTelemetry.Tests/Shared/InMemoryEventListener.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="InMemoryEventListener.cs" company="OpenTelemetry Authors">
+// <copyright file="InMemoryEventListener.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");